Fit a generalized linear model for differential expression

Fit a generalized linear model in order to model the expression of genes coming from a scRNA count matrix. The resulting object retains the inferred parameters that can later be tested in order to find differentially expressed genes.

fit_linear_model(
  input_matrix,
  model_matrix,
  ncounts,
  group_matrix = NULL,
  gene_specific_model_tensor = NULL,
  kernel_input = NULL,
  gene_names = NULL,
  cell_names = NULL,
  inference_method = "SVI",
  method_specific_args = list()
)

Arguments

input_matrix

Matrix of counts representing gene expression data for individual cells. Each row corresponds to a gene, and each column represents a single cell.

model_matrix

Matrix also known as design matrix, it represents the relationship between the response variable and the predictor variables in the model. Each row represents a cell and each columns represent a predictor variable (e.g experimental conditions, biological factors, treatment groups, batch effects, ...)

ncounts

Vector with length equal to the number of cells. It contains some kind of factor relate to the library size. For example, it might contain the total UMI of each cell (i.e. colSums(input_matrix)) or the UMI of each cell divided by the average across all of them.

gene_names

Vector containing the names of the genes

cell_names

Vector containing the names of the cells

inference_method

String. Either "SVI" or "HMC"

method_specific_args

List containing additional arguments. The available arguments differs between the inference algorithms.

SVI only:

optimizer_name optimizer, one of "ClippedAdam", "Adam", and "SGD";

steps number of iterations of the optimization algorithm;

lr learning rate for the optimize;

gamma_lr parameters to tune the decay of the learning rate using "ClippedAdam";

batch_size number of data points or observations sampled from the input matrix in each iteration of the optimization algorithm;

HMC only:

num_samples number of iterations after the warmup-phase, it also indicates the posterior samples each chain will produce;

num_chains number of chains for the optimization algorithm;

warmup_steps number of iterations of the warmup-phase;

Shared:

cuda ;

jit_compile ;

full_cov ;

theta_bounds ;

init_loc ;